requirements.txt
flask
flask-mysqldb
python-dotenv
google-generativeai
